Tomato Disease and Classification/Demo Web App/requirements.txt
streamlit==0.88.0
numpy==1.19.5
tensorflow-cpu==2.6.0
Pillow==8.3.2
